requirements.txt
xlrd==1.2.0
XlsxWriter==3.1.5
urllib3==2.0.5
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
tqdm==4.66.1
transformers==4.33.2
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
regex==2023.8.8
requests==2.31.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyparsing==3.1.1
pytest==7.4.3
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
olefile==0.46
openai==0.28.1
packaging==23.1
pandas==2.1.1
docx2txt==0.8
ebcdic==1.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Cors==4.0.0
aspose-words==23.9.0
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.4
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-search==1.0.0b2
azure-search-documents==11.4.0b8
azure-storage-blob==12.18.2
beautifulsoup4==4.8.2